Link shared gRPC libraries.
Guestlib is loaded before gRPC libraries are initialized when those
libraries are linked statically, which causes segfault at GrpcLibraryCodegen.
The problem is solved by linking gRPC libraries dynamically.

A side-effect of this change is that, for some reason, cmake fails to
specify the rpath for libgpr.so and libaddress_sorting.so in the gRPC
build directory. So for now the application should be run by:

```
LD_LIBRARY_PATH=PATH/TO/LIBGUESTLIB:PATH/TO/BUILD_DIR/third_party/grpc/lib ./app
```