Annotate cudaIpcCloseMemHandle as example.
When we get errors like "symbol cudaIpcCloseMemHandle not defined", we
know the API that we need to annotate. And here are the steps:

1. Copy the API signature from `/usr/local/cuda-10.1/include/cuda_runtime_api.h`
   to the spec (we can work on `pt_dump.c` first).
2. Annotate cudaIpcCloseMemHandle with Lapis.
3. Rebuild AvA and run the test program again. We get the next API that
   we need to annotate.
4. Repeat this process.

Note that we can use dummy annotations for the API at this point. See
`cudaEventCreateWithFlags` as an example in `pt_dump.c`: the API may not
be used at all, but putting a dummy annotation in the spec will allow us
to run the benchmark without linking errors. Once the API is called by
the benchmark, it will abort with `cudaEventCreateWithFlags is not implemented`.
At that time, we can start to write real annotations for the API.

Because building AvA takes quite some time, we may annotate a bunch of
APIs at each time:
Search `cudaIpcCloseMemHandle` in the PyTorch's source. We can find that
it is used in `c10/cuda/CUDACachingAllocator.cpp:1368`. Look at the CUDA
APIs used in that file, we know that we also need to annotate APIs such
as `cudaStreamGetCaptureInfo`, `cudaEventCreateWithFlags`, `cudaIpcOpenMemHandle`.